Enabled editing and implemented switching to the Settings screen though the info button. Also deleted old Settings button.
